requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.79.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
numpy==1.24.2
pandas==1.5.3
find / -name .git -exec dirname {} \; -prunepydantic==1.9.2
python-dotenv==0.20.0
pytorch_lightning==1.6.0
PyYAML==6.0
ray[tune]==1.11.0
scikit-learn==1.0.2
sentencepiece==0.1.96
setuptools==59.5.0
sniffio==1.2.0
starlette==0.19.1
torch==1.10.2
torchvision==0.11.3
transformers==4.17.0
typing-extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
openai==0.27.2
python-dotenv==0.21.1
openpyxl==3.1.2